requirements.txt
numpy
pandas
yfinance 
plotly
streamlit
azure.identity
azure.storage.blob